AimsApplyTransform: adjust the header transformations
The 'transformations' and 'referentials' header fields of the transformed data are now updated with respect to the applied transformation, so that they point to the same coordinate space as in the input file. If a non-affine transformation is applied, all header transformations are discarded.

These behaviours can be disabled by passing --keep-transforms.