Fix undefined behaviour in pyaims import
Sigh. Finally tracked down this issue which wasted my day. The undefined
behaviour occurred because npy_import_array was missing a return statement in
the success case. This is technically undefined behaviour, so the compiler is
allowed to do whatever it wants, although the return value is completely
ignored. CLang 6.0.0 feels like it is being clever and changes the code path in
this case, taking the error path, which raised the Python exception.

Sigh.

Fixes: https://github.com/brainvisa/aims-free/issues/27