moved the modified AimsLabels2Tex command to AimsVol2Tex
AimsVol2Tex used to be a modified clone of AimsLabels2Tex, and was
dealing with non-label (intensity) volumes. The new commands does both,
there is no need for two commands, and we could/did not maintain modifs
to both. The former AimsVol2Tex had "strange" options "parallel volume"
and "parallel texture" which have been removed (but I guess masking
and/or using the command twice will do the same).