added a reader for transformations in file headers (like nifti images)
the reader uses a "fake extension" .trmhdr, ex: image.nii.gz.trmhdr
The .trmhdr file does not exist, but the reader loads the image header
and extracts info from it. It may cause problems with workflow systems
with file transfers, to it's mabye not the ultimate solution...
Optins "index" or "target" allow to select the transformation in the
header (several ones can be present) either by index or by referential name.