implemented a Reader<Transformation3d> which can read Affine and FFD transforms
registered in the IO system. It is implemented in Soma-IO API for the
affine, and in aims API for the FFD (as each used to be implemented like
this), but they are available in both APIs for Transformation3d.

In python, use: aims.read('filename', dtype='Transformation3d'). Casts
are made on the returned object.