allow AimsGraphConvert to use GIFTI labels/colors tables in textures/images
If the conversion source is a volume with a .minf containing the
GIFTI_labels_table property, then it will be used automatically.
Otherwise an object (texture, volume) .minf file may be specified using
the --labels option: the minf format will be recognized, and any
GIFTI_labels_table in it will be used.

The new option -n allows to save a nomenclature (hierarchy) with the
labels table and colors.