worked aroud a very strange bug in bindings internals
in rc_ptr to its internal object conversion, calls to
PyObject_HasAttrString() (C for hasattr()) sometimes fail and return 0,
while re-calling them a second time immediately actually works.
It took me time to identify the problem, and the fix is a bad
work-around (calling the function twice) because I don't understand the
cause of the problem.